# src/main/resources/swagger/api.yaml
openapi: "3.0.0"
info:
  title: Backend API
  description: Backend API example.
  version: 1.0.0
paths:
  /employees/{employeeId}:
    get:
      summary: Info for a specific employee
      operationId: getEmployee
      tags:
        - employees
      parameters:
        - name: employeeId
          in: path
          description: Employee id
          required: true
          schema:
            type: integer
            format: int64
        - name: filterResigned
          in: query
          schema:
            type: boolean
      responses:
        200:
          description: Requested employee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        400:
          description: Missing/wrong parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /employees:
    get:
      summary: List all employees
      operationId: getEmployees
      tags:
        - employees
      parameters:
        - name: filterResigned
          in: query
          schema:
            type: boolean
      responses:
        200:
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        400:
          description: Missing/wrong parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create or update a employee
      operationId: createEmployees
      tags:
        - employees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        200:
          description: Updated employee
        400:
          description: Missing/wrong parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove all employees except authenticated user... May be dangerous.
      operationId: removeAllEmployeesExceptMe
      tags:
        - employees
      responses:
        200:
          description: Employees removed
        400:
          description: Missing/wrong parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /employees/top:
    get:
      summary: List top managers
      operationId: getTopManagers
      tags:
        - employees
      responses:
        200:
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        400:
          description: Missing/wrong parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /countries:
    get:
      summary: List all countries
      operationId: getCountries
      tags:
        - countries
      responses:
        200:
          description: List of countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
        400:
          description: Missing/wrong parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cache:
    delete:
      summary: Clear Spring caches
      operationId: clearCache
      tags:
        - cache
      responses:
        200:
          description: Cache cleaned
components:
  schemas:
    Employee:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        firstname:
          type: string
        lastname:
          type: string
        entryDate:
          type: string
          format: date-time
        salary:
          type: integer
          format: int32
        teamSize:
          type: integer
          format: int32
        country:
          $ref: '#/components/schemas/Country'
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'

    Country:
      type: object
      properties:
        code:
          type: string
        labelFr:
          type: string
        labelEn:
          type: string

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string